Add function declarations
Separated functions which find their target using a pointer to the node,
and functions which find their target using the node's index in the
list.

This minor version update came soon after the last. This is because most
of my recent work has been tweaking the API and this has caused the
version number to jump up quickly.

Hopefully the current interface is sufficiently robust to stop this from
continuing.