Rename header files and restructure inc directory
In doing this, I'm unsure of the best way to handle file version
numbers, specifically the version number of the log header.

I've renamed the log header and incremented its patch version number. I
don't want to reset the file to version 0.0.0 or bring it up to 0.2.0,
but I also don't feel that this is really a patch update.

Rather than worry about what version it should be, I'm going to make it
a patch and move on.